Reworked the management of query parameters in :index actions
The old code was working fine, but it was impossible to override the
query parameter settings in the controller, because the framework got
them directly from :params.
Additionally, it was not possible to use the
ActiveController::Parameters functionality to filter them.
The new code expects the query and pagination parameters to be passed
to the service object’s :index method. And that method was moved up
into the base class.