Enhancements to the asset-based ACL
Mostly added the :owner permission, which is granted by the owner of an asset to itself. This permission is used to filter assets accessible to a user in one SQL statement against the access grants table.

Added the .accessible_query class method in the access record class to get lists of assets to which a user has a given set of permissions.

Also fixed test scripts and a few bugs that the tests uncovered.